+++ R/nlmixr_vpc_theme.R
@@ -30,6 +30,3 @@ nlmixr_vpc_theme <- vpc::new_vpc_theme(list(obs_color = "#1F4E79",
                                             sim_median_linetype = "dashed",
                                             sim_median_size = 1,
                                             bin_separators_color = "#000000" ) )
-
-##'@export
-nlmixr::vpc