+++ NEWS.md
@@ -1,3 +1,7 @@
+# xpose.nlmixr 0.3.0
+
+* Move xpose.nlmixr to https://github.com/nlmixr2/xpose.nlmixr
+
 # xpose.nlmixr 0.2.0
 
 * Added a `NEWS.md` file to track changes to the package.